requirements.txt
streamlit
python-dotenv
openai
PyPDF2
SpeechRecognition
gTTS
audio-recorder-streamlit
reportlab
